Generations!
There is crossover (mating) and mutation as well as some random chance
for a stalk to be cloned into the next generation.

The fit population increases nicely over generations at the current
settings--from 179 to 779 in 50 generations!

Currently something is wrong with the starting X coordinate of each
stalk. Over the generations, they gradually cluster to the right side.